Load config from file with fallbacks
Removed some tests because I'm not sure how to test this without passing
the home dir in. I want to avoid passing it in so I can try several
paths in the future, and support Windows.

Also kind of wanted to use Results and bubble errors up to the top but
I'm sure there's a better way of handling that. Chosen to panic in
from_file instead because continuing when there's a config file but it's
not valid seems like odd behaviour.